Look for existing chromedriver in entire PATH, not just cwd
Without this change, a user who followed the README instructions and
now has chromedriver installed in, say, /usr/bin or /usr/local/bin
needs to change to that directory before running mintapi in order to
use that binary, which is non-intuitive and undocumented.

Additionally, if user doesn't have chromedriver installed, this change
downloads the binary to ~/.mintapi instead of current directory to
avoid littering unrelated directories with chromedriver binaries.